Fixed code

+++ x86_asm_uptm/array_reverser/array_reverser.cpp
@@ -3,12 +3,12 @@
 #include <array>
 #include <stdlib.h>
 
-extern "C" void reverse_array(int8_t* dst, const int8_t * src, int8_t n);
+extern "C" void reverse_array(int32_t* dst, const int32_t * src, int32_t n);
 
 int main()
 {
     constexpr size_t n = 10;
-    std::array<int8_t, n> src_array, dst_array;
+    std::array<int32_t, n> src_array, dst_array;
 
     for (size_t i = 0; i < n; i++) {
         src_array[i] = static_cast<int8_t>(i);
